Ensured all commands utilized `requires` to avoid subsystem conflicts
This is one of the reasons for why the lift freaked out when switching from Autonomous to Teleop. I believe that, in addition to all the static commands in `Robot.java`, two commands which operated the lift were running at the same time. One was an auto command, the other the teleop command. The teleop command set the motors to 0, while the auto command tried to move them. This back and forth fighting between the commands is most likely the cause of this issue, however to be on the safe side more oganization will need to occur so that is absolutely sure that this issue can never occur at competition.